Rename non-reusable SearchInput component to SearchBar
@brianlovin this component is very tighly tied to the Search screen, and
it doesn't really make sense to reuse any of it, so I renamed it
"SearchBar" to be more explicit about that and moved it to the Search view.
