+++ mobile/views/Search/SearchBar/index.js
