+++ mobile/views/Search/SearchBar/style.js
